requirements.txt
flask
flask-cors
flask-session
google-generativeai
PyPDF2
pandas
